[MojoLLDB] Perform a cast in formating to silence compiler warning, NFC (#12172)
Silences

```
/Users/abduld/code/modular/KGEN/lib/MojoLLDB/ExpressionParser/JITExecutionUnit.cpp:739:30: warning: format specifies type 'unsigned long long' but the argument has type 'unsigned long' [-Wformat]
            funcRange.first, funcRange.second);
            ~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~
/Users/abduld/code/modular/third-party/llvm-project/lldb/include/lldb/Utility/Log.h:348:27: note: expanded from macro 'LLDB_LOGF'
      log_private->Printf(__VA_ARGS__);                                        \
                          ^~~~~~~~~~~
1 warning generated.
```

MODULAR_ORIG_COMMIT_REV_ID: 506ab7e6f757306240025d708547f5d34640f7c1

+++ KGEN/lib/MojoLLDB/ExpressionParser/JITExecutionUnit.cpp
@@ -736,7 +736,7 @@ Status JITExecutionUnit::disassembleFunction(Stream &stream,
   }
 
   LLDB_LOGF(log, "Function's code range is [0x%" PRIx64 "+0x%" PRIx64 "]",
-            funcRange.first, funcRange.second);
+            funcRange.first, (unsigned long long)funcRange.second);
 
   Target *target = exeCtx.GetTargetPtr();
   if (!target) {
